backend/python/requirements.txt
# Simple Google Generative AI
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
